Update Jenkinsfile
combine nested stages with the when directive so that you don’t have repeat a when condition for every stage it applies. We will also add a set of pull request specific stages to include the Build and Push Container Image stage that is a required status check for merging pull requests to the main branch of you insurance-frontend repository.